requirements.txt
numpy>=1.20.0
matplotlib>=3.4.0
pillow>=8.2.0
torch>=1.10.0
transformers>=4.15.0
opencv-python>=4.5.3
huggingface-hub>=0.4.0
pyyaml>=6.0
requests>=2.25.1
tqdm>=4.62.0
scikit-learn>=1.0.0
pandas>=1.3.0
seaborn>=0.11.2
fastapi>=0.68.0
uvicorn>=0.15.0